services/company-info/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
httpx==0.23.0
python-dotenv==0.19.0 